Detect wordpress xmlrpc (sometimes get ssrf)
Provides a way to detect the wordpress xmlrpc endpoint that can help and leads to possible ssrf sometimes.
I am not sure if it is worth to add. But I have got a few good findings using this.

Let me know if I'm doing it right, You guys are doing great work, Love this project. Kudos..! 
Trying to contribute a little.